Update on Proposal draft
In this initial version, I utilized solely the 'IIB Data Breaches - LATEST - breaches.csv' from the dataset folder. I started experimenting with the data cleaning section and produced 'breaches_cleaned.csv.' Please note that the code provided in this first commit is intended for reference purposes and is not the final version.